LCMT v3.5.1: Create New Lights menu
This new release comes with a new light creation menu that is in it's
basic stages as I'm coding new stuff when I find the need to when using
LCMT in my current project.

It also provides crucial bug fixes for Correct Layer Creation of Vray
Lights and Mental Ray specific lights.

More to come soon!