Update lcmtv_current_release.py
Taken care of refixing Bug 1 and 8. No even if the "mental ray type" lights are hidden they are not counted as geometry which was causing the make layers procedures to crash. Also solved problems when Vray wasn't installed because it gave an error and caused the Bug fixes 1 and 8 to malfunction.