v. 3.6.4 Create Vray Lights
V. 3.6.3 had the ability to show on the lighting list only the lights that were on the current renderlayer the user was at. That functionality started breaking some others so it was discarded until better integration.
Updates that give you the opportunity to create VRay lights from inside LCMT. Only if Vray is installed.
Now renderlayers have renderlayername + _Light appended to the name for cleaner render layer naming.